refactor(web): extract FilesystemSelector component
Move the filesystem selector (dropdown / ReadOnlyField plus default and
keep-data hints) out of partition-form/FilesystemFields.tsx into a shared,
presentation-only component. Rename its partition-specific prop to the
generic selectedDevice so the logical volume and formattable device forms
can reuse it. Option building stays with the caller.
