+++ web/src/components/storage/partition-form/FilesystemFields.tsx
@@ -23,13 +23,11 @@
 import React, { useEffect, useMemo } from "react";
 import { unique, isEmpty } from "radashi";
 import { sprintf } from "sprintf-js";
-import { HelperText, HelperTextItem } from "@patternfly/react-core";
-import Text from "~/components/core/Text";
-import FieldNestedContent from "~/components/form/FieldNestedContent";
 import { withForm } from "~/hooks/form";
+import FilesystemSelector from "~/components/storage/shared/FilesystemSelector";
 import { defaultOptions, isReusingPartition, FILESYSTEM_TYPE, FILESYSTEM_ACTION } from "./fields";
 import { useVolumeTemplate } from "~/hooks/model/system/storage";
-import { deviceLabel, filesystemLabel, formattedPath } from "~/components/storage/utils";
+import { deviceLabel, filesystemLabel } from "~/components/storage/utils";
 import { _ } from "~/i18n";
 
 import type { Storage as System } from "~/model/system";
@@ -45,120 +43,6 @@ type FilesystemFieldsContentProps = {
   committedMountPoint: string;
   filesystem: string;
 };
-
-type FilesystemSelectorProps = {
-  form;
-  defaultFilesystem: ConfigModel.FilesystemType | undefined;
-  committedMountPoint: string;
-  filesystemOptions: Array<
-    { value: string; label: React.ReactNode; description?: React.ReactNode } | { divider: true }
-  >;
-  usableFilesystems: ConfigModel.FilesystemType[];
-  selectedPartition?: System.Device;
-  isFallback: boolean;
-};
-
-function keepFsText(filesystem) {
-  // TRANSLATORS: %s is a filesystem (eg. XFS)
-  return sprintf(_("%s (keep data)"), filesystemLabel(filesystem));
-}
-
-function defaultFsText(filesystem, mountPoint, isFallback) {
-  if (isFallback) {
-    return sprintf(
-      // TRANSLATORS: %s is a filesystem type (eg. XFS)
-      _("%s (default file system for generic partitions)"),
-      filesystemLabel(filesystem),
-    );
-  }
-
-  return sprintf(
-    // TRANSLATORS: %1$s is a filesystem (eg. XFS), %2$s is a mount point (eg. "/home")
-    _("%1$s (default file system for %2$s)"),
-    filesystemLabel(filesystem),
-    formattedPath(mountPoint),
-  );
-}
-
-/**
- * Filesystem selector: dropdown or ReadOnlyField depending on available options.
- *
- * When "Default" is selected, an inline hint explains which filesystem will be
- * used for the current mount point.
- *
- * When only one concrete filesystem type is available (excluding the "Default"
- * option), shows a ReadOnlyField with the type name instead of a dropdown. This
- * applies to mount points like swap and /boot/efi that constrain the filesystem
- * to a single type.
- *
- * When reusing a partition and the user selects a format option, shows a warning
- * that existing data will be destroyed.
- */
-function FilesystemSelector({
-  form,
-  defaultFilesystem,
-  committedMountPoint,
-  filesystemOptions,
-  usableFilesystems,
-  selectedPartition,
-  isFallback,
-}: FilesystemSelectorProps) {
-  const isSingleType = usableFilesystems.length === 1;
-
-  if (isSingleType && defaultFilesystem) {
-    return (
-      <form.AppField name="filesystem">
-        {(field) => (
-          <field.ReadOnlyField label={_("File system")} text={filesystemLabel(defaultFilesystem)} />
-        )}
-      </form.AppField>
-    );
-  }
-
-  return (
-    <form.AppField name="filesystem">
-      {(field) => (
-        <field.DropdownField label={_("File system")} options={filesystemOptions}>
-          {(value) => {
-            const currentFsType = selectedPartition?.filesystem?.type;
-            const isKeeping = value === FILESYSTEM_ACTION.REUSE && currentFsType;
-            const isAuto =
-              value === FILESYSTEM_TYPE.AUTO && !!defaultFilesystem && !!committedMountPoint;
-            const showHint = isKeeping || isAuto;
-            const showWarning = selectedPartition && !isKeeping;
-
-            if (!showHint && !showWarning) return null;
-
-            const fsText = isKeeping
-              ? keepFsText(currentFsType)
-              : defaultFsText(defaultFilesystem, committedMountPoint, isFallback);
-
-            return (
-              <FieldNestedContent>
-                {showHint && (
-                  <Text isBold textStyle="textColorSubtle">
-                    {fsText}
-                  </Text>
-                )}
-                {showWarning && (
-                  <HelperText>
-                    <HelperTextItem variant="warning">
-                      {sprintf(
-                        // TRANSLATORS: %s is partition name like "/dev/vdd2"
-                        _("Existing data on %s will be destroyed when installation begins."),
-                        deviceLabel(selectedPartition),
-                      )}
-                    </HelperTextItem>
-                  </HelperText>
-                )}
-              </FieldNestedContent>
-            );
-          }}
-        </field.DropdownField>
-      )}
-    </form.AppField>
-  );
-}
 
 /**
  * Inner component that renders filesystem fields and handles auto-reset logic.
@@ -271,7 +155,7 @@ const FilesystemFieldsContent = withForm({
           committedMountPoint={committedMountPoint}
           filesystemOptions={filesystemOptions}
           usableFilesystems={usableFilesystems}
-          selectedPartition={isReuse && canKeepCurrentFilesystem ? selectedPartition : undefined}
+          selectedDevice={isReuse && canKeepCurrentFilesystem ? selectedPartition : undefined}
           isFallback={isFallbackVolume}
         />
